Function to get prefixes announced per ASN added
Expecting to see a way to get the information per ASN
via the RIS whois protocol, I discovered this not the case.
As the RIPE RIS dashboard is providing the information,
the function is relying on the Ajax interface to gather the data.
I hope this is an acceptable use of the Ajax interface for RIPE RIS.
Maybe RIPE RIS will decide to add the ability to query an ASN
directly via the RIS whois later on. The Scrappy perl module is required
if you intend to use the function.